+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Bump datadog to v1.3.0
 - Update flipper to v0.141.0
 - Add a query to check whether the latest BP entry is high for the patient
+- Load clinical decision support info when the patient summary screen is created
 
 ### Features